pyproject.toml
[project]
name = "backend-gdtd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "alembic>=1.14.1",
    "asyncio>=4.0.0",
    "fastapi>=0.124.4",
    "psycopg[binary]>=3.2.13",
    "pydantic-settings>=2.8.1",
    "sqlmodel>=0.0.29",
    "uvicorn>=0.33.0",
]
